added funding badge
added a funding badge to our project to make it easier for people to support us financially while contributing to our project. this commit adds the `.github/funding.yml` file that contains the configuration for the badge.

this commit adds the `.github/funding.yml` file to the project, which contains the configuration for the funding badge. this badge will enable contributors to easily and securely support the project by making a donation. the badge will be visible on the project page and will link to our preferred donation platform.